when payment success then item delete

+++ src/Pages/DashBoard/AdminDashboard/ManageClassCard.jsx
@@ -1,6 +1,7 @@
 import Swal from "sweetalert2";
 import { AuthContext } from "../../../Provider/AuthProvider";
 import { useContext, useEffect, useState } from "react";
+import { Link } from "react-router-dom";
 
 
 
@@ -19,7 +20,7 @@ const ManageClassCard = ({ item }) => {
     const approveClass = item => {
         console.log(item);
         if (user && user.email) {
-            const cartItem = { name, image, instructor , seats ,enrolled,  price, email: user.email }
+            const cartItem = { name, image, instructor , seats ,enrolled, status:'approve' ,  price, email: user.email }
             fetch('http://localhost:5000/approveclasses', {
                 method: 'POST',
                 headers: {
@@ -40,6 +41,38 @@ const ManageClassCard = ({ item }) => {
                         })
                         // Update the item status
                         item.status = 'approved';
+                        setIsButtonDisabled(true);
+                        localStorage.setItem(`isButtonDisabled_${item._id}`, JSON.stringify(true));
+
+                    }
+                })
+        }
+
+    }
+    const deniedClass = item => {
+        console.log(item);
+        if (user && user.email) {
+            const cartItem = { name, image, instructor , seats ,enrolled, status:'denied' ,  price, email: user.email }
+            fetch('http://localhost:5000/approveclasses', {
+                method: 'POST',
+                headers: {
+                    'content-type': 'application/json'
+                },
+                body: JSON.stringify(cartItem)
+            })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.insertedId) {
+                        // refetch cart to update the number of items in the cart
+                        Swal.fire({
+                            position: 'top-center',
+                            icon: 'success',
+                            title: ' Added Class on the Class Page.',
+                            showConfirmButton: false,
+                            timer: 1500
+                        })
+                        // Update the item status
+                        item.status = 'Denied';
                         setIsButtonDisabled(true);
                         localStorage.setItem(`isButtonDisabled_${item._id}`, JSON.stringify(true));
 
@@ -69,8 +102,8 @@ const ManageClassCard = ({ item }) => {
             <td>{item.status}</td>
             <th className="space-y-2 flex flex-col">
                 <button onClick={() => approveClass(item)} className="btn btn-primary btn-xs" disabled={isButtonDisabled}>approved</button>
-                <button className="btn btn-warning btn-xs" disabled={isButtonDisabled}>denied</button>
-                <button className="btn btn-accent btn-xs">feedback</button>
+                <button onClick={() => deniedClass(item)} className="btn btn-warning btn-xs" disabled={isButtonDisabled}>denied</button>
+                <Link to='/dashboard/feedback'> <button className="btn btn-accent btn-xs">feedback</button></Link>
             </th>
         </tr>
     );
